requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.45.0
django-autoslug==1.9.8
django-autoslug-field==0.2.3
django-bootstrap-form==3.4
django-cors-headers==3.8.0
django-etc==1.3.1
django-filter==21.1
django-froala-editor==4.0.7
django-hitcount==1.3.3
django-quill-editor==0.1.22
django-ranged-response==0.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-simple-captcha==0.5.14
django-taggit==2.0.0
django-tinymce==3.4.0
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
drf-yasg==1.20.0
filelock==3.4.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
Pillow==8.3.2
platformdirs==2.4.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.4.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
turtle==0.0.1
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
